Add test that throws SymProblem error
Not sure if this is functioning as expected.
Bonds along a and b are sym. equiv. accroding to P 4 group but the lengths differ by d=0.5 Ang.
Here an error is thrown when d < tol but surely that is then consistent (i.e. tol is tolerance in bond lengths considered equiv.).
Also I think it should actually be tolDist (tolD variable in code) not tol used in this comparison.