Fix flaky MacOS test with mex
For some reason with a large omega_tol added to the diagonal,
on MacOS `chol` gives erroneous eigenvalues (of ~1e-3 instead of zero).
This error goes down with the value of omega_tol.
On other platforms the error is not so large (~1e-6).
In the non-mex code, it tries first with smaller omega_tol,
(==`min(eig(ham))`) so doesn't get this error but the mex code
just uses `omega_tol` input directly.